backend/twitter_followers/requirements.txt
appdirs==1.4.3
attrs==18.2.0
backcall==0.1.0
black==18.9b0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.2
Django==2.1.7
django-environ==0.4.5
django-extensions==2.1.5
django-model-utils==3.1.2
djangorestframework==3.9.1
future==0.17.1
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
oauthlib==3.0.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.8
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
python-twitter==3.5
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
toml==0.10.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
